Refacor: Split up league models into different files
